# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baraka API",
        "description": "API backend pour la plateforme Baraka de microcrédit",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/baraka-backend/public",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/simulations": {
            "post": {
                "tags": [
                    "Simulations"
                ],
                "summary": "Créer une simulation de crédit",
                "operationId": "d1b3e086a3bd50cd913d747e739befab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant_emprunte",
                                    "duree_mois",
                                    "taux_utilise"
                                ],
                                "properties": {
                                    "montant_emprunte": {
                                        "type": "integer",
                                        "example": 500000
                                    },
                                    "duree_mois": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "taux_utilise": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "id_offre": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "frais_dossier": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "assurance": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Simulation créée"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/v1/simulations/preview": {
            "post": {
                "tags": [
                    "Simulations"
                ],
                "summary": "Calculer une simulation sans persistance",
                "operationId": "0af4bc185526050775406d091e6044c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant_emprunte",
                                    "duree_mois",
                                    "taux_utilise"
                                ],
                                "properties": {
                                    "montant_emprunte": {
                                        "type": "integer",
                                        "example": 500000
                                    },
                                    "duree_mois": {
                                        "type": "integer",
                                        "example": 24
                                    },
                                    "taux_utilise": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "id_offre": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "frais_dossier": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "assurance": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Résultat de simulation (non persisté)"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/v1/simulations/{id}": {
            "get": {
                "tags": [
                    "Simulations"
                ],
                "summary": "Afficher une simulation",
                "operationId": "8357c7c1ce068f73fb326f691d64033f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la simulation"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Simulation introuvable"
                    }
                }
            }
        },
        "/api/v1/users/me/simulations": {
            "get": {
                "tags": [
                    "Simulations"
                ],
                "summary": "Lister les simulations de l'utilisateur authentifié",
                "operationId": "08bab9084ca692095f815bc3d1f19729",
                "responses": {
                    "200": {
                        "description": "Liste paginée des simulations"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT obtenu via POST /api/auth/otp/verify"
            }
        }
    },
...
}